[#122] Fixed tests
