+++ development/src/Index.ts
@@ -1,3 +1,23 @@
 import { game } from './Game';
+import { onActionTriggered, onFellOffTheBoardTriggered, renderScreen } from './LogicAdapter';
+import { highlightSquare, initializeBoards } from './ui/BoardManager';
+import { initializeEventListeners, setOnAction, setOnFellOffTheBoard, setOnHighlight } from './ui/Events';
+import { Logger } from './ui/Logger';
 
-game.initializeGame();
+function setGameEventHandlers() {
+  setOnAction(onActionTriggered);
+  setOnFellOffTheBoard(onFellOffTheBoardTriggered);
+  setOnHighlight(highlightSquare);
+}
+
+function initializeUI() {
+  initializeBoards();
+  initializeEventListeners();
+  renderScreen();
+  setGameEventHandlers();
+}
+
+initializeUI();
+game.initialize();
+
+Logger.logGeneral('Game started!');
