+++ development/src/logic/items/Coin.ts
@@ -2,7 +2,7 @@ import { Item } from './Items';
 import { Piece } from '../pieces/Pieces';
 import { Logger } from '../../ui/Logger';
 import { coinResource } from '../../ui/Resources';
-import { Position } from '../pieces/PiecesHelpers';
+import { Position } from '../pieces/PiecesUtilities';
 
 export class Coin extends Item {
   constructor(position: Position) {
