+++ development/src/logic/items/Trap.ts
@@ -2,7 +2,7 @@ import { Item } from './Items';
 import { Piece } from '../pieces/Pieces';
 import { Logger } from '../../ui/Logger';
 import { trapResource } from '../../ui/Resources';
-import { Position } from '../pieces/PiecesHelpers';
+import { Position } from '../pieces/PiecesUtilities';
 import { spawnItemOnBoard } from '../../LogicAdapter';
 import { game } from '../../Game';
