+++ development/src/logic/pieces/Bishop.ts
@@ -1,7 +1,7 @@
 import { bishopResource } from '../../ui/Resources';
 import { Piece } from './Pieces';
 import { Player, PlayerColors } from '../Players';
-import { Position, Square, simulateMove } from './PiecesHelpers';
+import { Position, Square, simulateMove } from './PiecesUtilities';
 
 export class Bishop extends Piece {
   constructor(position: Position, player: Player) {
