+++ development/src/logic/pieces/Pawn.test.ts
@@ -1,9 +1,11 @@
+import { game } from '../../Game';
 import { OVERWORLD_BOARD_ID } from '../Constants';
 import { Player, PlayerColors } from '../Players';
 import { Pawn } from './Pawn';
-import { Position } from './PiecesHelpers';
+import { Position } from './PiecesUtilities';
 
-jest.mock('../../LogicAdapter.ts');
+jest.mock('../../ui/BoardManager.ts', () => ({}));
+jest.mock('../../ui/Screen.ts', () => ({}));
 
 const whitePlayer = new Player(PlayerColors.WHITE);
 const blackPlayer = new Player(PlayerColors.BLACK);
@@ -15,6 +17,7 @@ describe('Piece movements', () => {
       boardId: OVERWORLD_BOARD_ID,
     };
     const pawn = new Pawn(initialPosition, whitePlayer);
+    game.setPieces([pawn]);
 
     const singleStepMove: Position = {
       coordinates: [0, 5],
