+++ development/src/logic/pieces/Rook.test.ts
@@ -1,9 +1,11 @@
+import { game } from '../../Game';
 import { OVERWORLD_BOARD_ID } from '../Constants';
 import { Player, PlayerColors } from '../Players';
-import { Position } from './PiecesHelpers';
+import { Position } from './PiecesUtilities';
 import { Rook } from './Rook';
 
-jest.mock('../../LogicAdapter.ts');
+jest.mock('../../ui/BoardManager.ts', () => ({}));
+jest.mock('../../ui/Screen.ts', () => ({}));
 
 const whitePlayer = new Player(PlayerColors.WHITE);
 
@@ -14,6 +16,7 @@ describe('Piece movements', () => {
       boardId: OVERWORLD_BOARD_ID,
     };
     const rook = new Rook(initialPosition, whitePlayer);
+    game.setPieces([rook]);
 
     const newPosition: Position = {
       coordinates: [0, 2],
