+++ development/src/logic/pieces/Rook.ts
@@ -1,7 +1,7 @@
 import { rookResource } from '../../ui/Resources';
 import { Piece } from './Pieces';
 import { Player, PlayerColors } from '../Players';
-import { Position, Square, simulateMove } from './PiecesHelpers';
+import { Position, Square, simulateMove } from './PiecesUtilities';
 
 export class Rook extends Piece {
   constructor(position: Position, player: Player) {
