+++ package.json
@@ -25,21 +25,18 @@
     "eslint": "~8.46.0",
     "eslint-config-prettier": "8.1.0",
     "html-webpack-plugin": "^5.6.0",
-    "jest": "29.7.0",
+    "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.4.1",
     "nx": "17.2.7",
     "nx-cloud": "^16.0.4",
     "parcel": "^2.10.3",
     "prettier": "^3.1.1",
-    "ts-jest": "^29.1.0",
+    "ts-jest": "^29.1.1",
     "ts-loader": "^9.5.1",
     "ts-node": "10.9.1",
     "typescript": "~5.1.3",
     "webpack": "^5.89.0",
     "webpack-cli": "^5.1.4",
     "webpack-dev-server": "^4.15.1"
-  },
-  "dependencies": {
-    "tslib": "^2.3.0"
   }
 }
